Added vue toast
