+++ package.json
@@ -17,6 +17,7 @@
     "vue": "^2.6.14",
     "vue-hacktiv8-footer": "^0.0.9",
     "vue-router": "^3.5.1",
+    "vue-toast-notification": "^0.6.2",
     "vue-web-cam": "^1.9.0",
     "vuex": "^3.6.2"
   },
