+++ src/components/ProductCard.vue
@@ -27,6 +27,9 @@ export default {
     methods: {
         addToCart(productPackage) {
             this.$store.commit('addToCart', productPackage)
+            this.$toast.success('Item added to your cart!', {
+                position: 'top-right'
+            });
         }
     },
     computed: {
