+++ src/main.js
@@ -3,10 +3,12 @@ import App from './App.vue'
 import router from './router'
 import store from './store'
 import { BootstrapVue} from 'bootstrap-vue'
-// Import Bootstrap an BootstrapVue CSS files (order is important)
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
-// Make BootstrapVue available throughout your project
+import VueToast from 'vue-toast-notification';
+import 'vue-toast-notification/dist/theme-sugar.css';
+
+Vue.use(VueToast);
 Vue.use(BootstrapVue, {
   BModal: {
     'hide-footer': true,
